+++ packages/liveblocks-core/src/lib/stringify.ts
@@ -21,3 +21,19 @@ function replacer(_key: string, value: unknown) {
 export function stableStringify(value: unknown): string {
   return JSON.stringify(value, replacer);
 }
+
+/**
+ * Drop-in replacement for JSON.stringify(), which will log any payload to the
+ * console if it could not be stringified somehow.
+ */
+export function stringifyOrLog(value: unknown): string {
+  try {
+    return JSON.stringify(value);
+  } catch (err) {
+    /* eslint-disable rulesdir/console-must-be-fancy */
+    console.error(`Could not stringify: ${(err as Error).message}`);
+    console.error(value);
+    /* eslint-enable rulesdir/console-must-be-fancy */
+    throw err;
+  }
+}