+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## vNEXT (not yet published)
 
+### `@liveblocks/client`
+
+- Log more details in specific error cases to help debugging
+
 ### `@liveblocks/react`
 
 - Increases the allowed stale time for polled user threads data. Only affects